+++ ui/svg/src/element/render_context.cpp
@@ -4,20 +4,20 @@ float RenderContext::ResolveX(ui::Length& x) {
   if (x.unit == ui::Length::Unit::Percentage) {
     return view_box.width * x.value / 100;
   }
-  return x.value;
+  return (float)x.value;
 }
 float RenderContext::ResolveY(ui::Length& y) {
   if (y.unit == ui::Length::Unit::Percentage) {
     return view_box.height * y.value / 100;
   }
-  return y.value;
+  return (float)y.value;
 }
 
 float RenderContext::ResolveRX(ui::Length& rx) {
-  int r = ResolveX(rx);
+  float r = ResolveX(rx);
   return std::min((float)r, view_box.width / 2);
 }
 float RenderContext::ResolveRY(ui::Length& ry) {
-  int r = ResolveY(ry);
+  float r = ResolveY(ry);
   return std::min((float)r, view_box.height / 2);
 }
